requirements/production.txt
# Production requirements for Django-Twilio-Call
# Include base requirements
-r base.txt

# Production WSGI server
gunicorn==21.2.0
gevent==23.9.1

# Production database connections
psycopg2-binary==2.9.9

# Production caching
django-redis==5.4.0
hiredis==2.2.3

# Static file serving
whitenoise[brotli]==6.6.0

# Security enhancements
django-security==0.17.0
django-cors-headers==4.3.1

# Monitoring and observability
sentry-sdk[django]==1.38.0
django-prometheus==2.3.1
prometheus-client==0.19.0

# Health checks
django-health-check==3.17.0

# Performance optimizations
django-querycount==0.8.3
django-silk==5.0.4  # Only for staging

# AWS/Cloud storage
boto3==1.34.23
django-storages[boto3]==1.14.2

# Background task monitoring
flower==2.0.1

# SSL/TLS support
certifi==2023.11.17

# Production logging
structlog==23.2.0
django-structlog==4.1.0

# Rate limiting
django-ratelimit==4.1.0

# Environment configuration
python-decouple==3.8
django-environ==0.11.2

# Database connection pooling
django-db-connection-pool[postgresql]==1.2.5

# Memory profiling (optional)
memory-profiler==0.61.0
psutil==5.9.6

# Process management
supervisor==4.2.5